chore(node/engine): small touchups to the engine
