+++ crates/node/engine/src/state/mod.rs
@@ -3,5 +3,8 @@
 mod core;
 pub use core::EngineState;
 
+mod error;
+pub use error::EngineStateBuilderError;
+
 mod builder;
-pub use builder::{EngineStateBuilder, EngineStateBuilderError};
+pub use builder::EngineStateBuilder;
